requirements.txt
discord.py
pickle-mixin
requests